fixed step numbers and bullet appearance
I couldn't figure out the correct syntax to get the list items to be at level 2 to be indented under step 3 of "Running Single Inspections" so both are currently level 1 (solid bullet) so that they are obviously of the same level rather than one being empty and one filled. I use InDesign and FrameMaker normally and need to figure out Asciidoc . Sorry.