Update nbm-filetemplates.adoc
Added the `:registration:` tag with today's date and fixed the numbering issues. Apparently, AsciiDoc does not abide by the `[start=#]` option if there is a newline immediately following the number:

```
[start=3]
1.
Text here...
```
So, I removed those newlines so that the text for the numbered point is on the same line as the number. Every appears correct in the preview now.

-SC